packages/neural-intelligence/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neural-intelligence"
version = "0.1.0"
description = "Decentralized Superintelligent Agent System for Commerce"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "NinjaPay Team", email = "team@ninjapay.xyz"}
]
keywords = ["ai", "agents", "fintech", "fetch.ai", "singularitynet", "metta", "uagents"]

dependencies = [
    "uagents>=0.15.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "websockets>=11.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "redis>=5.0.0",
    "psycopg2-binary>=2.9.0",
    "asyncpg>=0.28.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "hyperon>=0.1.0",  # MeTTa runtime
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "ruff>=0.0.285",
    "mypy>=1.5.0",
]

[tool.setuptools]
packages = ["agents", "api", "knowledge", "sdk"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true
